third_party/blink/web_tests/external/wpt/css/css-color/parsing/relative-color-out-of-gamut-expected.txt

This is a testharness.js-based test.
[FAIL] Property color value 'rgb(from color(display-p3 0 1 0) r g b / alpha)'
  Colors do not match.\nActual:   color(srgb -0.511814 1.01832 -0.310726)\nExpected: color(srgb -0.5116 1.01827 -0.31067).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected -0.5116 +/- 0.0001, expected -0.5116 but got -0.511814
[FAIL] Property color value 'rgb(from lab(100 104.3 -50.9) r g b)'
  Colors do not match.\nActual:   color(srgb 1.59343 0.58802 1.40564)\nExpected: color(srgb 1.5935 0.58776 1.40555).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.58776 +/- 0.0001, expected 0.58776 but got 0.58802
[FAIL] Property color value 'rgb(from lch(100 116 334) r g b)'
  Colors do not match.\nActual:   color(srgb 1.59328 0.588284 1.40527)\nExpected: color(srgb 1.59336 0.58802 1.40517).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.58802 +/- 0.0001, expected 0.58802 but got 0.588284
[FAIL] Property color value 'rgb(from oklab(1 0.365 -0.16) r g b)'
  Colors do not match.\nActual:   color(srgb 1.59295 0.360371 1.38571)\nExpected: color(srgb 1.59297 0.3602 1.38555).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.3602 +/- 0.0001, expected 0.3602 but got 0.360371
[FAIL] Property color value 'rgb(from oklch(1 0.399 336.3) r g b)'
  Colors do not match.\nActual:   color(srgb 1.59328 0.358736 1.38663)\nExpected: color(srgb 1.59331 0.35856 1.38647).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.35856 +/- 0.0001, expected 0.35856 but got 0.358736
[FAIL] Property color value 'hsl(from color(display-p3 0 1 0) h s l / alpha)'
  Colors do not match.\nActual:   color(srgb -0.511814 1.01832 -0.310726)\nExpected: color(srgb -0.5116 1.01827 -0.31067).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected -0.5116 +/- 0.0001, expected -0.5116 but got -0.511814
[FAIL] Property color value 'hsl(from lab(100 104.3 -50.9) h s l)'
  Colors do not match.\nActual:   color(srgb 1.59343 0.58802 1.40564)\nExpected: color(srgb 1.5935 0.58776 1.40555).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.58776 +/- 0.0001, expected 0.58776 but got 0.58802
[FAIL] Property color value 'hsl(from lch(100 116 334) h s l)'
  Colors do not match.\nActual:   color(srgb 1.59328 0.588284 1.40527)\nExpected: color(srgb 1.59336 0.58802 1.40517).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.58802 +/- 0.0001, expected 0.58802 but got 0.588284
[FAIL] Property color value 'hsl(from oklab(1 0.365 -0.16) h s l)'
  Colors do not match.\nActual:   color(srgb 1.59295 0.360371 1.38571)\nExpected: color(srgb 1.59297 0.3602 1.38555).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.3602 +/- 0.0001, expected 0.3602 but got 0.360371
[FAIL] Property color value 'hsl(from oklch(1 0.399 336.3) h s l)'
  Colors do not match.\nActual:   color(srgb 1.59328 0.358734 1.38664)\nExpected: color(srgb 1.59331 0.35856 1.38647).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.35856 +/- 0.0001, expected 0.35856 but got 0.358734
[FAIL] Property color value 'hwb(from color(display-p3 0 1 0) h w b / alpha)'
  Colors do not match.\nActual:   color(srgb -0.511814 1.01832 -0.310726)\nExpected: color(srgb -0.5116 1.01827 -0.31067).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 0, expected -0.5116 +/- 0.0001, expected -0.5116 but got -0.511814
[FAIL] Property color value 'hwb(from lab(100 104.3 -50.9) h w b)'
  Colors do not match.\nActual:   color(srgb 1.59343 0.58802 1.40564)\nExpected: color(srgb 1.5935 0.58776 1.40555).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.58776 +/- 0.0001, expected 0.58776 but got 0.58802
[FAIL] Property color value 'hwb(from lch(100 116 334) h w b)'
  Colors do not match.\nActual:   color(srgb 1.59328 0.588284 1.40527)\nExpected: color(srgb 1.59336 0.58802 1.40517).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.58802 +/- 0.0001, expected 0.58802 but got 0.588284
[FAIL] Property color value 'hwb(from oklab(1 0.365 -0.16) h w b)'
  Colors do not match.\nActual:   color(srgb 1.59295 0.360371 1.38571)\nExpected: color(srgb 1.59297 0.3602 1.38555).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.3602 +/- 0.0001, expected 0.3602 but got 0.360371
[FAIL] Property color value 'hwb(from oklch(1 0.399 336.3) h w b)'
  Colors do not match.\nActual:   color(srgb 1.59328 0.358736 1.38664)\nExpected: color(srgb 1.59331 0.35856 1.38647).\nError: assert_array_approx_equals: Numeric parameters are approximately equal. property 1, expected 0.35856 +/- 0.0001, expected 0.35856 but got 0.358736
Harness: the test ran to completion.
